new converters for thickness (margin)
